ele lana a exceção, mas ainda n tem o catch apropriado quando é lançado de uma função filha.

+++ include/instructions/instruction_set/misc.h
@@ -416,7 +416,7 @@ class NewArray : public Instruction {
       std::stringstream *ss, const bool &wide, int *pc,
       const int &delta_tab) override {
     unsigned char atype = *++*code_it;
-    (*ss) << Opcodes::getMnemonic(this->opcode) << " " << atype << " "
+    (*ss) << Opcodes::getMnemonic(this->opcode) << " " << static_cast<int>(atype) << " "
           << this->getType(atype);
     *delta_code = 1;
     return {};
